Temp state with bug by use of Transformations.switchMap() for update of var asteroids in repository
I tried to use Transformations.switchMap() to update var asteroids in Repository.
App stops now with white screen.
For test I used for init and for update the same most simple call to database.
Because both will lead to same result in asteroids var we will not be able to see a difference on screen for now.